requirements.txt
greenlet==2.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
protobuf==3.20.1
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0